[update] プロフィールの変更に伴ってサイドバー等のユーザー情報を更新するための処理
