+++ pages/_app.js
@@ -2,11 +2,8 @@ import '../styles/globals.css'
 import AuthProvider from '../components/AuthContext'
 
 function MyApp({ Component, pageProps }) {
-	return (
-		<AuthProvider>
-			<Component {...pageProps} />
-		</AuthProvider>
-	)
+	const getLayout = Component.getLayout || (page => page)
+	return <AuthProvider>{getLayout(<Component {...pageProps} />)}</AuthProvider>
 }
 
 export default MyApp
