+++ styles/components/layout.module.css
@@ -1,0 +1,5 @@
+.container {
+	min-height: 100vh;
+	background: #59bc6c;
+	display: flex;
+}
