+++ styles/users/Home.module.css
@@ -109,96 +109,3 @@
 .tab__content__container {
 	padding: 1rem 2rem;
 }
-
-/* _______ output _______ */
-.content__output {
-	width: 100%;
-	height: 226px;
-	background: white;
-	border-radius: 10px;
-	padding: 1rem;
-	margin-bottom: 2rem;
-	box-shadow: 0px 4px 4px 1px #000;
-	position: relative;
-	overflow: hidden;
-}
-
-.user__container {
-	display: flex;
-	flex-direction: row;
-	align-items: center;
-}
-
-.user__image__area {
-	display: flex;
-	justify-content: center;
-	width: 70px;
-}
-
-.user__image {
-	border-radius: 9999px;
-}
-
-.user__info {
-	margin-left: 1rem;
-}
-
-.user__name {
-	font-size: 10px;
-}
-
-.output__words {
-	font-size: 15px;
-}
-
-/* book */
-.book__container {
-	display: flex;
-	flex-direction: row;
-}
-
-.book__image {
-	border-radius: 20px;
-}
-
-.book__info {
-	margin-left: 1rem;
-}
-
-/* _______ edit and delete button ______ */
-.btn__group {
-	position: absolute;
-	top: 0;
-	right: 0;
-	display: flex;
-	flex-direction: column;
-	gap: 1rem;
-	padding: 1rem;
-}
-
-.btn__crud {
-	border-radius: 20px;
-	padding: 0.3rem 1.2rem;
-	background: white;
-}
-
-/* _______ like button ______ */
-.btn__container {
-	position: absolute;
-	bottom: 10px;
-	left: 130px;
-	width: 60%;
-	display: flex;
-	flex-direction: row;
-	justify-content: space-evenly;
-}
-
-.btn {
-	display: flex;
-	flex-direction: row;
-	align-items: center;
-}
-
-.count {
-	margin-left: 1rem;
-}
